feat: :construction: A potensial way of handeling errors
Made a simple way to showchase that an error has occured. The user noe recives a text on the screen where the image would apire if everything was ok. Pls take a look and give feedback. NB the API gives error code 500 for every error and not 400 as the backend said?

#11